+++ gatspy/__init__.py
@@ -7,7 +7,7 @@ from __future__ import absolute_import
 from . import datasets, periodic
 
 __all__ = ['datasets', 'periodic']
-__version__ = '0.3.dev0'
+__version__ = '0.3'
 
 
 def setup():