+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Fix bug in ``predict()`` for ``LombScargle`` methods when ``center_data=False``
 - Fix bug in ``score()`` for ``LombScargle`` methods when ``center_data=False`` and ``fit_offset=False``
+- Make ``supersmoother`` a soft dependency
 
 ## (v0.2.1) bugfix release (19 Aug 2015)