+++ CHANGES.md
@@ -8,6 +8,7 @@
 - New feature: ``gatspy.periodic.TrendedLombScargle`` extends ``LombScargle``
   by adding a linear trend parameter ``d*t`` to the fitted model
 - Fixed download URL for RR Lyrae datasets
+- Dropped support for Python 2.6 and 3.3
 
 ## (v0.2.1) bugfix release (19 Aug 2015)